+++ packages/snap/src/permissions.ts
@@ -60,4 +60,24 @@ export const originPermissions = new Map<string, string[]>([
       InternalMethod.IsSynchronousMode,
     ],
   ],
+  [
+    'metamask.github.io',
+    [
+      // Keyring methods
+      KeyringRpcMethod.ListAccounts,
+      KeyringRpcMethod.GetAccount,
+      KeyringRpcMethod.CreateAccount,
+      KeyringRpcMethod.FilterAccountChains,
+      KeyringRpcMethod.UpdateAccount,
+      KeyringRpcMethod.DeleteAccount,
+      KeyringRpcMethod.ExportAccount,
+      KeyringRpcMethod.ListRequests,
+      KeyringRpcMethod.GetRequest,
+      KeyringRpcMethod.ApproveRequest,
+      KeyringRpcMethod.RejectRequest,
+      // Custom methods
+      InternalMethod.ToggleSyncApprovals,
+      InternalMethod.IsSynchronousMode,
+    ],
+  ],
 ]);